Struct blockchain::blockchain::Blockchain

pub struct Blockchain {
    tip: Arc<Reference>,
    heads: LruCache<HeaderId, Arc<Reference>>,
    cache: LruCache<HeaderId, Arc<Reference>>,
}

Fields
tip: Arc<Reference>heads: LruCache<HeaderId, Arc<Reference>>cache: LruCache<HeaderId, Arc<Reference>>
Implementations

impl Blockchain

pub fn new(configuration: &Configuration, block0: Arc<Reference>) -> Self

start a blockchain with the given block as starting point

pub fn tip(&self) -> Arc<Reference>

pub fn branches(&self) -> Iter<'_, HeaderId, Arc<Reference>>

get an iterator for all the branches currently being considered by the Blockchain.

The tip is already included in the list too and it may be that the some branches in the list are no longer Head only.

pub fn put(&mut self, block: &Block) -> Result<Event, Error>

fn put_head( &mut self, reference: Arc<Reference>, new_branch: bool, epoch_transition: bool ) -> Event