Struct blockchain::miner::Miner

pub struct Miner {
    pub id: u32,
    pub network: HashSet<(u32, String)>,
    pub blocks: Vec<Block>,
    pub sockip: String,
    pub wallets: HashSet<(u32, String)>,
    pub payload: Vec<String>,
    pub current_block_id: u32,
}

This struct is used to represent a Miner in our Blockchain Network *id - the ID of our Miner, should be unique within the network *network - a HashSet containing the IDs and IP addresses of every Miner inside the network *blocks - a Vector containing the Blockchain *sockip - the IP address on which our Miner will listen for incoming transactions *wallets - a HashSet containing the IDs and IP addresses of every wallets binded to it *payload - TBD *current_block_id - TBD

Fields
id: u32network: HashSet<(u32, String)>blocks: Vec<Block>sockip: Stringwallets: HashSet<(u32, String)>payload: Vec<String>current_block_id: u32
Implementations

impl Miner

pub fn new(id: u32, socket: String) -> Self

CONSTRUCTOR socket - an ip:port string representing where is the Miner listening Returns a new Miner with a TcpListener that listens to the given ip:port

pub fn get_id(&self) -> u32

Getter for the id attribute of a Miner

pub fn display_network(&self)

Function used to print the network attribute of a Miner

pub fn join(&self, destination: String)

Function to join an existing network

destination - the ip:port of the Miner we want to join

pub fn send_message(
&self,
destination: &String,
message: &String,
flag: Flag
) -> Result<u8, &'static str>

Function to send a message

stream - Tcp Stream.

message - The message to send.

pub fn broadcast_to_network(&self, message: &String, flag: Flag, sender: String)

This function is used by a Miner to Broadcast an incoming message to the whole network *message - the message to be broadcasted *flag - the flag associated with the message *sender - the IP address of the miner which sended the original message

pub fn broadcast_threaded(&self, message: &String)

Message propagation to all neighbors

message - Message sent. Unused

pub fn retrieve_next_id(&self) -> u32

Used by an existing Miner when a new miner joins the network and asks for it's unique ID Returns an integer which will be assigned as the joining Miner's ID

pub fn retrieve_next_wallet_id(&self) -> u32

Used by an existing Miner when a new wallet is binded to it and asks for it's unique ID Returns an integer which will be assigned as the new wallet's ID

pub fn handle_client(&mut self, stream: TcpStream)

This function handles the incoming requests/messages sended by other entities *stream - a TCPStream instance which contains the data that was received The function decodes the incoming message, then act accordingly

pub fn add_to_network(&mut self, peer_id: u32, peer_addr: String) -> bool

Function to add a Miner to the network peer_id - an integer to identify the Miner, should be unique in the network peer_addr - the socket on which the Miner is listening, should be unique aswell Update the current Miner's network, returns true if the Miner was added to the newtork, false if the Miner was already in the network

pub fn remove_from_network(&mut self, peer_id: u32, peer_addr: String) -> bool

Function to remove a Miner from the network peer_id - an integer to identify the Miner peed_addr - the socket of the Miner we want to remove from the network Update the current Miner's network, returns true if the Miner was deleted from the newtork, false if the Miner wasn't in the network

pub fn listen(self)

Function to listen for incoming Streams from the network Read the stream and spawn a thread to handle the received data

pub fn ask_for_id(socket: &String, destination: &String) -> u32

Used by Miners to ask the Miner to which it is connecting to give us our ID *socket - our IP address *destination - the IP address of the Miner we are asking for an ID

pub fn ask_miner_for_wallet_id(socket: &String, miner: &String) -> u32

Used by wallets to ask the Miner to which it is connecting to give us our ID *socket - our IP address *destination - the IP address of the Miner we are asking for an ID

pub fn add_to_wallets(&mut self, peer_id: u32, peer_addr: String) -> bool

Simple function to add a new wallet to our Miner's wallet list *peer_id - the ID of the new wallet *peer_addr - the IP address of the new wallet

pub fn refresh_nodes_status(&mut self)

Function to refresh all nodes status and remove those are not accessible

pub fn health_check(
&mut self,
destination: &String,
id: &u32
) -> Result<u8, &'static str>

health_check ping the destination. If it doesn't respond ok or at time, removing the destination node from network

pub fn check_block(&self, block: Block) -> bool

Function to check if the received block is valid

*block the received block Return true if the block is valid, false else

pub fn hash_block(&self, transactions: String) -> Block

Function to mine block *transaction List of transactions to mine *MINING_DIFFICULTY use Return the mined block