Struct blockchain::miner::wallet::Wallet

pub struct Wallet {
    pub id: u32,
    pub miner: String,
    pub socket: String,
}

This struct represent the wallets *id - an integer than should be unique to each wallet within a miner's wallet list *miner - the IP address of the miner to which the wallet is binded *socket - the IP address on which the wallet listens for incoming messages

Fields
id: u32miner: Stringsocket: String
Implementations

impl Wallet

pub fn new(socket: String, miner: String, id: u32) -> Self

CTOR *socket - the IP address where the wallet listens *miner - the IP address miner to which that wallet is tied *id - the unique ID that must be assigned to this wallet Returns a new wallet with parameters initialized as given

pub fn listen_for_user_input(&self)

This function allows our wallet to listen to the commands given by the user on std input It checks for validity of the input, and act accordingly

pub fn handle_user_input(
&self,
command: UserCommand,
target: String,
message: String
) -> String

Performs the action which the user gave as input

pub fn handle_message(&self, stream: TcpStream) -> String

This handles incoming message, by decoding them and transforming them into usable data *stream - is a TcpStream instance containing the bytes that we received

pub fn send_message(
&self,
destination: &String,
message: &String,
flag: Flag
) -> Result<u8, &'static str>

Function to send a message to another entity on the network

stream - Tcp Stream.

message - The message to send.