Struct blockchain::Block

pub struct Block {
    num: i32,
    data: String,
    prevHash: u64,
}

The 'Block' struct.

Fields
num: i32

The block number.

data: String

The actual data of the block. Currently can only be a 'String', future work will generalize this to arbitrary data.

prevHash: u64

The hash of the previous 'Block' in the 'Chain'. If this is the first block, 'prevHash' is set to 0.

Methods

impl Block

pub fn new(n: i32, d: String, prevh: u64) -> Block

Creates a new block.

Packages a block number, string data, and hash of previous block into a new block. This method is used primarily for testing purposes.

pub fn get_hash(&self) -> u64

Returns the hash of a block.

This method contains the basic sequence of calls required to get the actual hash of a struct. It is used when inserting new blocks into a chain, or when validating a chain.