Struct blockchain::Chain

pub struct Chain {
    lst: VecDeque<Block>,
}

Putting the 'Chain' in blockchain.

Fields
lst: VecDeque<Block>

The only element of this struct, a vecdeque of 'Block's.

Methods

impl Chain

pub fn new() -> Chain

Returns a new (empty) blockchain.

pub fn insert_block(&mut self, dataStr: String)

Inserts a new block into the chain given the data for the new block.

If this function is called on an empty chain, the genesis block is inserted. Otherwise, we calculate required fields for the new block and push it to the back of the vecdeque.

pub fn validate_chain(&mut self) -> bool

Determines whether a chain is valid.

A chain is valid if the 'prevHash' value of each block matches the hash of the previous block. If this condition holds for all blocks in the chain, then it is valid.